% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency.matrix.R
\name{adjacency.matrix}
\alias{adjacency.matrix}
\title{Compute an Adjacency Matrix}
\usage{
adjacency.matrix(partition)
}
\arguments{
\item{partition}{An integer vector of cluster labels for \eqn{n} items. Two
items are in the same subset (i.e., cluster) if their labels are equal.}
}
\value{
An adjacency matrix, i.e., binary matrix whose \eqn{(i,j)} element is
  one if and only if \eqn{i} and \eqn{j} have the same cluster label in the
  supplied partition.
}
\description{
Given partition, this function computes an adjacency matrix, i.e., binary
matrix whose \eqn{(i,j)} element is one if and only if elements \eqn{i} and
\eqn{j} in the partition have the same cluster label.
}
\examples{
adjacency.matrix(iris.clusterings[1,])

}
\seealso{
\code{\link{psm}}
}
