% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.dendrogram.R
\name{as.dendrogram.salso.estimate}
\alias{as.dendrogram.salso.estimate}
\title{Dendrogram Based on Expected Partition Loss}
\usage{
\method{as.dendrogram}{salso.estimate}(object, ...)
}
\arguments{
\item{object}{An object of class \code{salso.estimate} obtained from the
\code{\link{salso}} function.}

\item{...}{Currently ignored.}
}
\value{
An object of class \code{salso.dendrogram}, which can be plotted with
  the \code{\link{plot}} function.
}
\description{
A dendrogram based on expected partition loss shows the relationships among
clusters when merging pairs of clusters such that the increase in the
expectation of the posterior loss is minimized.
}
\examples{
draws <- iris.clusterings
estimate <- salso(draws, loss="VI", nRuns=4, parallel=FALSE)
plot(as.dendrogram(estimate))

}
