% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonicalize_cluster_labels.R
\name{canonicalize_cluster_labels}
\alias{canonicalize_cluster_labels}
\title{Canonicalize Cluster Labels}
\usage{
canonicalize_cluster_labels(x)
}
\arguments{
\item{x}{A vector of cluster labels.}
}
\value{
A numeric vector containing cluster labels in cannonical form.
}
\description{
Canonicalize Cluster Labels
}
