% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlso.R
\name{dlso}
\alias{dlso}
\title{Draws-Based Latent Structure Optimization}
\usage{
dlso(psm, loss = c("binder", "pear", "VI.lb")[3], draws, parallel = FALSE)
}
\arguments{
\item{psm}{A pairwise similarity matrix, i.e., \eqn{n}-by-\eqn{n} symmetric
matrix whose \eqn{(i,j)} element gives the (estimated) probability that
items \eqn{i} and \eqn{j} are in the same subset (i.e., cluster) of a
partition (i.e., clustering).  If not provided, this argument is computed
from the \code{draws} argument.}

\item{loss}{One of \code{"binder"}, \code{"pear"}, or \code{"VI.lb"}.  See
\code{\link{partition.expected.loss}} for details on these loss functions.}

\item{draws}{A \eqn{B}-by-\eqn{n} matrix, where each of the \eqn{B} rows
represents a clustering of \eqn{n} items using cluster labels. For
clustering \eqn{b}, items \eqn{i} and \eqn{j} are in the same cluster if
\code{x[b,i] == x[b,j]}.}

\item{parallel}{Should the search use all CPU cores?  (Currently ignored
since parallelization is not implemented.)}
}
\value{
A list of the following elements: \describe{ \item{estimate}{An
  integer vector giving a partition encoded using cluster labels.}
  \item{loss}{A character vector equal to the \code{loss} argument.}
  \item{expectedLoss}{A numeric vector of length one giving the expected
  loss.} }
}
\description{
This function provides a partition to summarize a partition distribution
based a pairwise similarity matrix using the draws latent structure
optimization (DLSO) method, which is also known as the least-squares
clustering method (Dahl 2006). The method seeks to minimize an estimation
criterion by picking the minimizer among the partitions supplied by the
\code{draws} argument. The implementation currently supports the minimization
of three partition estimation criteria: "binder", "pear", and "VI.lb". For
details on these criteria, see \code{\link{partition.expected.loss}}.
}
\examples{
dlso(draws=iris.clusterings, loss="binder")
dlso(draws=iris.clusterings, loss="pear")
dlso(draws=iris.clusterings, loss="VI.lb")

}
\seealso{
\code{\link{partition.expected.loss}}, \code{\link{psm}},
  \code{\link{confidence}}, \code{\link{salso}}
}
