% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumerate.partitions.R
\name{enumerate.partitions}
\alias{enumerate.partitions}
\title{Enumerate Partitions of a Set}
\usage{
enumerate.partitions(nItems)
}
\arguments{
\item{nItems}{The size of the set \code{\{0, 1, ..., n-1\}}, i.e., \code{n}.}
}
\value{
A matrix of integers, where each row is a partition encoded as a
  vector of cluster labels.
}
\description{
This function produces a matrix whose rows provide all possible partitions of
the set of integers \code{\{0, 1, ..., n-1\}}. These partitions are provided
as cluster labels, where two items are in the same subset (i.e., cluster) if
their labels are equal.
}
\examples{
enumerate.partitions(5)

}
