% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iris.clusterings-data.R
\docType{data}
\name{iris.clusterings}
\alias{iris.clusterings}
\title{Clusterings of the Iris Data}
\format{
A 1000-by-150 matrix of 1000 randomly generated clusterings of the
  150 observations in the \code{iris} dataset.
}
\source{
Unknown.
}
\usage{
iris.clusterings
}
\description{
Randomly generated clusterings for the \code{iris} dataset.
}
\seealso{
\code{\link[datasets]{iris}}
}
\keyword{datasets}
