% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.salso.confidence.R
\name{plot.salso.summary}
\alias{plot.salso.summary}
\title{Confidence and Exemplar Plotting}
\usage{
\method{plot}{salso.summary}(x, data = NULL, showLabels = length(x$estimate) <= 50, ...)
}
\arguments{
\item{x}{An object returned by the \code{\link{confidence}} function.}

\item{data}{The data from which the distances were computed.}

\item{showLabels}{Should the names of items be shown in the plot?}

\item{...}{Currently ignored.}
}
\value{
\code{NULL}, invisibly.
}
\description{
This function produces confidence plots (e.g., heatmaps of pairwise
allocation probabilities) and exemplar plots. The "exemplar" refers to the
best representative of a particular cluster. See \code{\link{confidence}} for
further explanation.
}
\examples{
# For examples, use 'parallel=FALSE' per CRAN rules but, in practice, omit this.
draws <- iris.clusterings
est <- salso(draws, parallel=FALSE)
probs <- psm(draws, parallel=FALSE)
conf <- confidence(est, probs)
plot(conf)
plot(conf, data=iris)

}
\seealso{
\code{\link{confidence}}, \code{\link{psm}}, \code{\link{dlso}},
  \code{\link{salso}}
}
\author{
David B. Dahl \email{dahl@stat.byu.edu}
}
