% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psm.R
\name{psm}
\alias{psm}
\title{Compute the Pairwise Similarity Matrix}
\usage{
psm(x, parallel = TRUE)
}
\arguments{
\item{x}{A \eqn{B}-by-\eqn{n} matrix, where each of the \eqn{B} rows
represents a clustering of \eqn{n} items using cluster labels. For
clustering \eqn{b}, items \eqn{i} and \eqn{j} are in the same cluster if
\code{x[b,i] == x[b,j]}.}

\item{parallel}{Should the computation use all CPU cores?}
}
\value{
A \eqn{n}-by-\eqn{n} symmetric matrix whose \eqn{(i,j)} element
  gives the estimated expected number of times that items \eqn{i} and
  \eqn{j} are in the same subset (i.e., cluster or feature) based on the
  frequencies from the supplied clusterings or feature allocations.
}
\description{
Based on the provided samples, this function computes the \eqn{n}-by-\eqn{n}
matrix whose \eqn{(i,j)} element gives the (estimated) probability that items
\eqn{i} and \eqn{j} are in the same subset (i.e., cluster).  This is the mean
of the adjacency matrices of the provided samples.
}
\examples{
dim(iris.clusterings)
# Use 'parallel=FALSE' per CRAN rules for examples but, in practice, omit this.
probs <- psm(iris.clusterings, parallel=FALSE)
dim(probs)
probs[1:6, 1:6]

}
\seealso{
\code{\link{adjacency.matrix}}
}
