% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.salso.estimate.R
\name{summary.salso.estimate}
\alias{summary.salso.estimate}
\title{Summary of Partitions Estimated Using Posterior Expected Loss}
\usage{
\method{summary}{salso.estimate}(object, ...)
}
\arguments{
\item{object}{An object returned by the \code{\link{salso}} function.}

\item{...}{Currently ignored}
}
\value{
A list containing the score for each observation, exemplar
  observation for each cluster, a dendrogram object, and a vector for
  ordering observations in the heatmap plot.
}
\description{
Summary of Partitions Estimated Using Posterior Expected Loss
}
\examples{
# For examples, use 'parallel=FALSE' per CRAN rules but, in practice, omit this.
draws <- iris.clusterings
est <- salso(draws, parallel=FALSE)
summ <- summary(est)
plot(summ, type="heatmap")
plot(summ, type="dendrogram")
plot(summ, type="exemplar", data=iris)

}
