% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{binder}
\alias{binder}
\alias{lpear}
\alias{VI.lb}
\title{Compute a Partition Loss Function}
\usage{
binder(partitions, psm)

lpear(partitions, psm)

VI.lb(partitions, psm)
}
\arguments{
\item{partitions}{An integer matrix of cluster labels, where each row is a
partition given as cluster labels. Two items are in the same subset (i.e.,
cluster) if their labels are equal.}

\item{psm}{A pairwise similarity matrix, i.e., \code{n}-by-\code{n} symmetric
matrix whose \code{(i,j)} element gives the (estimated) probability that
items \code{i} and \code{j} are in the same subset (i.e., cluster) of a
partition (i.e., clustering).}
}
\value{
A numeric vector of length equal to the number of rows of
  \code{partitions}, where each element gives the value of the loss function.
}
\description{
Based on the supplied pairwise similarity matrix, these functions compute,
for the given partitions, 1. the expectation of the Binder loss
(\code{binder}), 2. the loss for the adjusted Rand index with the posterior
expected clustering (\code{lpear}), and 3. the lower bound of the expectation
of the variation of information loss (\code{VI.lb}).
}
\examples{
probs <- psm(iris.clusterings, parallel=FALSE)
binder(iris.clusterings[1:5,], probs)
lpear(iris.clusterings[1:5,], probs)
VI.lb(iris.clusterings[1:5,], probs)

}
