% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence.R
\name{confidence}
\alias{confidence}
\title{Compute Clustering Confidence}
\usage{
confidence(estimate, psm)
}
\arguments{
\item{estimate}{A vector of length \code{n}, where \code{i} and \code{j} are
in the same cluster if and only if \code{clustering[i] == clustering[j]}.}

\item{psm}{A pairwise similarity matrix, i.e., \code{n}-by-\code{n} symmetric
matrix whose \code{(i,j)} element gives the (estimated) probability that
items \code{i} and \code{j} are in the same subset (i.e., cluster) of a
partition (i.e., clustering).}
}
\value{
A list of the following elements: \describe{ \item{estimate}{The
  value of the \code{estimate} argument.}
  \item{psm}{The value of the \code{psm} argument.}
  \item{confidence}{asdf}
  \item{confidenceMatrix}{asdf}
  \item{exemplar}{asdf}
  \item{order}{asdf}
  }
}
\description{
This function computes the confidence values for \code{n} observations based
on a clustering estimate and the expected pairwise allocation matrix.
}
\examples{
probs <- psm(iris.clusterings, parallel=FALSE)
est <- salso(probs, nPermutations=50, parallel=FALSE)$estimate
conf <- confidence(est,probs)
conf

}
\seealso{
\code{\link{salso}}, \code{\link{dlso}}, \code{\link{psm}}
}
\author{
David B. Dahl \email{dahl@stat.byu.edu}
}
