% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlso.R
\name{dlso}
\alias{dlso}
\title{Draws-Based Latent Structure Optimization}
\usage{
dlso(candidates, loss = c("binder", "pear", "VI.lb", "VI")[3], x = NULL)
}
\arguments{
\item{candidates}{A \eqn{B}-by-\eqn{n} matrix, where each of the \eqn{B} rows
represents a clustering of \eqn{n} items using cluster labels. For
clustering \eqn{b}, items \eqn{i} and \eqn{j} are in the same cluster if
\code{x[b,i] == x[b,j]}.  One of the rows will be used as the paritition
estimate.}

\item{loss}{One of \code{"binder"}, \code{"pear"}, or \code{"VI.lb"}.  See
\code{\link{partition.loss}} for details on these loss functions.}

\item{x}{Either: 1. A \eqn{B}-by-\eqn{n} matrix, where each of the \eqn{B}
rows represents a clustering of \eqn{n} items using cluster labels. (For
clustering \eqn{b}, items \eqn{i} and \eqn{j} are in the same cluster if
\code{x[b,i] == x[b,j]}.), or 2. A pairwise similarity matrix, i.e.,
\eqn{n}-by-\eqn{n} symmetric matrix whose \eqn{(i,j)} element gives the
(estimated) probability that items \eqn{i} and \eqn{j} are in the same
subset (i.e., cluster) of a partition (i.e., clustering).  If \code{NULL},
this argument is set equal to \code{candidates}.}
}
\value{
A list of the following elements: \describe{ \item{estimate}{An
  integer vector giving a partition encoded using cluster labels.}
  \item{loss}{A character vector equal to the \code{loss} argument.}
  \item{expectedLoss}{A numeric vector of length one giving the expected
  loss.} }
}
\description{
This function provides a partition to summarize a partition distribution
using the draws latent structure optimization (DLSO) method, which is also
known as the least-squares clustering method (Dahl 2006). The method seeks to
minimize an estimation criterion by picking the minimizer among the
partitions supplied by the \code{draws} argument. The implementation
currently supports the minimization of the following partition estimation
criteria: "binder", "pear", "VI.lb", and "VI". For details on these criteria,
see \code{\link{partition.loss}}.
}
\details{
This function provides a partition to summarize a partition distribution
using the sequentially-allocated latent structure optimization (SALSO)
method. The implementation currently supports the minimization of the
following partition estimation criteria: "binder", "pear", "VI.lb", and "VI".
For details on these criteria, see \code{\link{partition.loss}}.
}
\examples{
# For examples, use 'parallel=FALSE' per CRAN rules but, in practice, omit this.
probs <- psm(iris.clusterings, parallel=FALSE)
dlso(iris.clusterings, loss="binder", x=probs)
dlso(iris.clusterings, loss="pear", x=probs)
dlso(iris.clusterings, loss="VI.lb")  # The psm will be computed if not provided.
dlso(iris.clusterings, loss="VI")
dlso(iris.clusterings[1:10,], loss="VI", x=iris.clusterings)  # Candidates can be constrained.

}
\seealso{
\code{\link{partition.loss}}, \code{\link{psm}},
  \code{\link{confidence}}, \code{\link{salso}}
}
