% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumerate.partitions.R
\name{enumerate.partitions}
\alias{enumerate.partitions}
\title{Enumerate Partitions of a Set}
\usage{
enumerate.partitions(nItems)
}
\arguments{
\item{nItems}{The size of the set \eqn{{1, 2, ..., n}}, i.e., \eqn{n}.}
}
\value{
A matrix of integers, where each row is a partition encoded as a
  vector of cluster labels.
}
\description{
This function produces a matrix whose rows provide all possible partitions of
the set \eqn{{1, 2, ..., n}}. These partitions are provided as cluster
labels, where two items are in the same subset (i.e., cluster) if their
labels are equal.
}
\examples{
# R_CARGO \dontrun{
# R_CARGO # Example disabled since Cargo was not found when installing from source package.
# R_CARGO # You can still run the example if you install Cargo. Hint: cargo::install().
enumerate.partitions(5)
# R_CARGO }

}
