% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumerate.permutations.R
\name{enumerate.permutations}
\alias{enumerate.permutations}
\title{Enumerate Permutations of Items}
\usage{
enumerate.permutations(nItems)
}
\arguments{
\item{nItems}{The size of the set \code{\{0, 1, ..., n-1\}}, i.e., \code{n}.}
}
\value{
A matrix of integers, where each row is a permutation.
}
\description{
This function produces a matrix whose rows provide all possible permutations of
the set of integers \code{\{0, 1, ..., n-1\}}.
}
\examples{
enumerate.permutations(5)

}
