% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.loss.R
\name{partition.loss}
\alias{partition.loss}
\alias{randi}
\alias{arandi}
\alias{vi.dist}
\title{Compute a Partition Loss}
\usage{
partition.loss(partition1, partition2, loss = c("binder", "pear", "VI.lb")[3])

randi(partition1, partition2)

arandi(partition1, partition2)

vi.dist(partition1, partition2)
}
\arguments{
\item{partition1}{An integer vector of cluster labels for \eqn{n} items. Two
items are in the same subset (i.e., cluster) if their labels are equal.}

\item{partition2}{An integer vector of cluster labels having the same length
as \code{partition1}.}

\item{loss}{One of \code{"binder"}, \code{"pear"}, or \code{"VI.lb"}.  See
\code{\link{partition.expected.loss}} for details on these loss functions.}
}
\value{
A numeric scalar giving the loss.
}
\description{
Given partitions \eqn{\pi*} and \eqn{\pi}, \code{\link{partition.loss}}
computes the loss when using \eqn{\pi*} to estimate \eqn{\pi}, based on one
of three partition losses. Smaller values of a loss indicate higher
concordance between the partitions. The implementation currently supports the
computation of the following partition loss functions: "binder", "pear", and
"VI.lb". For details on these criteria, see
\code{\link{partition.expected.loss}}.
}
\details{
The functions \code{\link{randi}}, \code{\link{arandi}}, and
\code{\link{vi.dist}} are convenience functions. Note that:
\itemize{
\item \code{partition.loss(p1, p2, "binder") = choose(length(p1), 2) * ( 1 - randi(p1, p2) )}
\item \code{partition.loss(p1, p2, "pear") = 1 - arandi(p1, p2)}
\item \code{partition.loss(p1, p2, "VI.lb") = vi.dist(p1, p2)}
}
}
\examples{
p1 <- iris.clusterings[1,]
p2 <- iris.clusterings[2,]

partition.loss(p1, p2, loss="binder")
partition.loss(p1, p2, loss="pear")
partition.loss(p1, p2, loss="VI.lb")

all.equal(partition.loss(p1, p2, "binder"), choose(length(p1), 2) * ( 1 - randi(p1, p2) ))
all.equal(partition.loss(p1, p2, "pear"), 1 - arandi(p1, p2))
all.equal(partition.loss(p1, p2, "VI.lb"), vi.dist(p1, p2))

}
\seealso{
\code{\link{partition.expected.loss}}
}
