% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.salso.confidence.R
\name{plot.salso.confidence}
\alias{plot.salso.confidence}
\title{Confidence and Exemplar Plotting}
\usage{
\method{plot}{salso.confidence}(x, estimate = NULL, data = NULL,
  show.labels = length(x$estimate) <= 50, ...)
}
\arguments{
\item{x}{An object returned by the \code{\link{confidence}} function.}

\item{estimate}{A vector of cluster labels, or \code{NULL}.}

\item{data}{The data from which the distances were computed.}

\item{show.labels}{Show the items names be shown in the plot?}

\item{...}{Currently ignored.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Functions to produce confidence plots (e.g., heatmaps of pairwise allocation probabilities) and exemplar plots.
}
\examples{
probs <- psm(iris.clusterings, parallel=FALSE)
est <- salso(probs, nPermutations=50, parallel=FALSE)$estimate
conf <- confidence(est, probs)
plot(conf)
plot(conf,data=iris)

}
\seealso{
\code{\link{psm}}, \code{\link{dlso}}, \code{\link{salso}}
}
\author{
David B. Dahl \email{dahl@stat.byu.edu}
}
