% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.salso.summary.R
\name{plot.salso.summary}
\alias{plot.salso.summary}
\title{Heatmap, Dendrogram, and Exemplar Plotting for Partition Estimation}
\usage{
\method{plot}{salso.summary}(
  x,
  type = c("heatmap", "dendrogram", "exemplar")[1],
  data = NULL,
  showLabels = length(x$estimate) <= 50,
  ...
)
}
\arguments{
\item{x}{An object returned by \code{summary(x)}, where \code{x} itself is returned by the \code{\link{salso}} function.}

\item{type}{A string equal to \code{"heatmap"}, \code{"dendrogram"}, or \code{"exemplar"}.}

\item{data}{The data from which the partition estimation was obtained.  This is required when \code{type='exemplar'} and ignored otherwise.}

\item{showLabels}{Should the names of items be shown in the plot when \code{type="heatmap"}?}

\item{...}{Currently ignored.}
}
\value{
\code{NULL}, invisibly.
}
\description{
This function produces one of three plots:  1. A heatmap shows the pairwise
allocation probabilities that items are clustered.
2. A dendrogram based on expected partition loss shows the relationships among
clusters when merging pairs of clusters such that the increase in the
expectation of the posterior loss is minimized.  3. An exemplar plot shows
pairs plots of all the variables with the exemplar (i.e., the most representative observation) of each cluster emphasized.
}
\examples{
# For examples, use 'parallel=FALSE' per CRAN rules but, in practice, omit this.
draws <- iris.clusterings
est <- salso(draws, parallel=FALSE)
summ <- summary(est)
plot(summ, type="heatmap")
plot(summ, type="dendrogram")
plot(summ, type="exemplar", data=iris)

}
\seealso{
\code{\link{salso}}, \code{\link{summary.salso.estimate}}
}
\author{
David B. Dahl \email{dahl@stat.byu.edu}
}
