% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxido.R
\docType{package}
\name{salso-package}
\alias{salso-package}
\title{salso: Search Algorithms and Loss Functions for Bayesian Clustering}
\description{
The SALSO algorithm is an efficient randomized greedy search method to find a point estimate for a random partition based on a loss function and posterior Monte Carlo samples. The algorithm is implemented for many loss functions, including the Binder loss and a generalization of the variation of information loss, both of which allow for unequal weights on the two types of clustering mistakes. Efficient implementations are also provided for Monte Carlo estimation of the posterior expected loss of a given clustering estimate. See Dahl, Johnson, Müller (2022) \doi{10.1080/10618600.2022.2069779}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dbdahl/salso}
  \item Report bugs at \url{https://github.com/dbdahl/salso/issues}
}

}
\author{
\strong{Maintainer}: David B. Dahl \email{dahl@stat.byu.edu} (\href{https://orcid.org/0000-0002-8173-1547}{ORCID})

Authors:
\itemize{
  \item Devin J. Johnson \email{devin.j.johnson7@gmail.com} (\href{https://orcid.org/0000-0003-2619-6649}{ORCID})
  \item Peter Müller \email{pmueller@math.utexas.edu}
}

Other contributors:
\itemize{
  \item Alex Crichton \email{alex@alexcrichton.com} (Rust crates: proc-macro2, cfg-if.) [copyright holder]
  \item "bluss" Ulrik Sverdrup (Rust crate: rawpointer, matrixmultiply, scopeguard, ndarray.) [copyright holder]
  \item Brendan Zabarauskas \email{bjzaba@yahoo.com.au} (Rust crate: approx.) [copyright holder]
  \item David B. Dahl \email{dahl@stat.byu.edu} (Rust crates: roxido, roxido_macro, dahl-partition, dahl-salso, dahl-bellnumber.) [copyright holder]
  \item David Tolnay \email{dtolnay@gmail.com} (Rust crates: syn, proc-macro2, quote, unicode-ident.) [copyright holder]
  \item DutchGhost (Rust crate: matrixmultiply.) [copyright holder]
  \item Enthought, Inc. (Rust crate: ndarray.) [copyright holder]
  \item Gilad Naaman \email{gilad.naaman@gmail.com} (Rust crate: memoffset.) [copyright holder]
  \item Jim Turner (Rust crate: ndarray.) [copyright holder]
  \item Josh Stone \email{cuviper@gmail.com} (Rust crate: autocfg.) [copyright holder]
  \item Melissa O'Neill (Rust crate: rand_pcg.) [copyright holder]
  \item Paul Dicker (Rust crate: rand_pcg.) [copyright holder]
  \item PCG Project contributors (Rust crate: rand_pcg.) [copyright holder]
  \item Ralf Jung (Rust crate: memoffset.) [copyright holder]
  \item rawpointer developers (Rust crate: rawpointer.) [copyright holder]
  \item R. Janis Goldschmidt (Rust crate: matrixmultiply.) [copyright holder]
  \item SciPy Developers (Rust crate: ndarray.) [copyright holder]
  \item Sean McArthur \email{sean@seanmonstar.com} (Rust crate: num_cpus.) [copyright holder]
  \item Stefan Lankes (Rust crate: hermit-abi.) [copyright holder]
  \item The Cranelift Project Developers (Rust crate: wasi.) [copyright holder]
  \item The Crossbeam Project Developers (Rust crates: crossbeam, crossbeam-channel, crossbeam-deque, crossbeam-epoch, crossbeam-queue, crossbeam-utils.) [copyright holder]
  \item The CryptoCorrosion Contributors (Rust crates: ppv-lite86, rand_chacha.) [copyright holder]
  \item The Go Authors (Rust crate: crossbeam-channel.) [copyright holder]
  \item The PCG Project Contributors (Rust crate: rand_pcg.) [copyright holder]
  \item The matrixmultiply Authors (Rust crate: matrixmultiply.) [copyright holder]
  \item The ndarray Developers (Rust crate: ndarray.) [copyright holder]
  \item The Rand Project Developers (Rust crates: rand_core, rand_chacha, rand_pcg, rand, getrandom.) [copyright holder]
  \item The Rust Project Developers (Rust crates: crossbmean-channel, rand_chacha, num-integer, rand_core, num-complex, libc, num-traits, rand.) [copyright holder]
  \item The scopeguard Developers (Rust crates: scopeguard.) [copyright holder]
  \item Unicode, Inc. (Rust crate: unicode-ident.) [copyright holder]
}

}
\keyword{internal}
