% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salso.R
\name{salso}
\alias{salso}
\title{Sequentially-Allocated Latent Structure Optimization}
\usage{
salso(psm, loss = c("VI.lb", "binder")[1], maxSize = 0, maxScans = 5,
  nPermutations = 5000, seconds = 10, parallel = TRUE)
}
\arguments{
\item{psm}{A pairwise similarity matrix, i.e., \code{n}-by-\code{n} symmetric
matrix whose \code{(i,j)} element gives the (estimated) probability that
items \code{i} and \code{j} are in the same subset (i.e., cluster) of a
partition (i.e., clustering).}

\item{loss}{Either \code{"VI.lb"} or \code{"binder"}, to indicate the desired
loss function.}

\item{maxSize}{Either zero or a positive integer.  If a positive integer, the
optimization is constrained to produce solutions whose number of subsets
(i.e., clusters) is no more than the supplied value.  If zero, the size is
not constrained.}

\item{maxScans}{The maximum number of reallocation scans after the intial
allocation.  The actual number of scans may be less than \code{maxScans}
since the method stops if the result does not change between scans.}

\item{nPermutations}{The desired number of permutations to consider when
searching for the minimizer.}

\item{seconds}{A time threshold in seconds after which the function will
return early (with a warning) instead of finishing all the desired
permutations.  Note that the function could take considerably longer,
however, because this threshold is only checked after each permutation is
completed.}

\item{parallel}{Should the search use all CPU cores?}
}
\value{
A list of the following elements: \describe{ \item{estimate}{An
  integer vector giving a partition encoded using cluster labels.}
  \item{loss}{A character vector equal to the \code{loss} argument.}
  \item{expectedLoss}{A numeric vector of length one giving the expected
  loss.} \item{nScans}{An integer vector giving the number of scans used to
  arrive at the supplied estimate.} \item{nPermutations}{An integer vector
  giving the number of permutations actually performed.} }
}
\description{
This function provides a point estimate for a partition distribution using
the sequentially-allocated latent structure optimization (SALSO) method. The
method seeks to minimize the Binder loss or the lower bound of the variation
of information loss. The SALSO method was presented at the workshop "Bayesian
Nonparametric Inference: Dependence Structures and their Applications" in
Oaxaca, Mexico on December 6, 2017. See
<https://www.birs.ca/events/2017/5-day-workshops/17w5060/schedule>.
}
\examples{
probs <- psm(iris.clusterings, parallel=FALSE)
salso(probs, nPermutations=50, parallel=FALSE)

}
